move fxn definitions out of init.py and rename change fxn names so do not have duplicate getFlowBySector() and getFlowByActivity()
