+++ examples/get_flows_by_activity.py
@@ -22,21 +22,22 @@ Examples of use of flowsa. Read parquet files as dataframes.
 
 """
 
-import flowsa
+from flowsa.common import seeAvailableFlowByModels
+import flowsa.flowbyactivity
 from flowsa.settings import fbaoutputpath
 
 # see all datasources and years available in flowsa
-flowsa.seeAvailableFlowByModels('FBA')
+seeAvailableFlowByModels('FBA')
 
 # Load all information for EIA MECS Land
-fba_mecs = flowsa.getFlowByActivity(datasource="EIA_MECS_Land", year=2014)
+fba_mecs = flowsa.flowbyactivity.getFlowByActivity(datasource="EIA_MECS_Land", year=2014)
 
 # only load state level water data and save as csv
-fba_usgs = flowsa.getFlowByActivity(datasource="USGS_NWIS_WU",
-                                    year=2015,
-                                    flowclass='Water',
-                                    geographic_level='state'
-                                    )
+fba_usgs = flowsa.flowbyactivity.getFlowByActivity(datasource="USGS_NWIS_WU",
+                                                   year=2015,
+                                                   flowclass='Water',
+                                                   geographic_level='state'
+                                                   )
 
 # save output to csv, maintain leading 0s in location col
 fba_usgs.Location = fba_usgs.Location.apply('="{}"'.format)
