+++ flowsa/flowbyactivity.py
@@ -17,12 +17,17 @@ from functools import partial, reduce
 from typing import Literal, List
 import fedelemflowlist
 import pandas as pd
-from flowsa import settings, metadata, log, geo, validation, naics, common, \
+
+import flowsa.exceptions
+from flowsa import settings, metadata, geo, validation, naics, common, \
     sectormapping, generateflowbyactivity
+from flowsa.flowsa_log import log
+from flowsa.settings import DEFAULT_DOWNLOAD_IF_MISSING
+from flowsa.flowbyfunctions import filter_by_geoscale
 from flowsa.flowby import _FlowBy, flowby_config, NAME_SEP_CHAR
 
 if TYPE_CHECKING:
-    from flowsa import FlowBySector
+    from flowsa.flowbysector import FlowBySector
 
 class FlowByActivity(_FlowBy):
     _metadata = [*_FlowBy()._metadata]
@@ -71,7 +76,7 @@ class FlowByActivity(_FlowBy):
         return _FBASeries
 
     @classmethod
-    def getFlowByActivity(
+    def return_FBA(
         cls,
         full_name: str,
         year: int = None,
@@ -677,7 +682,7 @@ class FlowByActivity(_FlowBy):
             download_sources_ok: bool = True
             ) -> 'FlowBySector':
 
-        from flowsa import FlowBySector
+        from flowsa.flowbysector import FlowBySector
 
         if 'activity_sets' in self.config:
             try:
@@ -854,3 +859,40 @@ class _FBASeries(pd.Series):
     @property
     def _constructor_expanddim(self) -> 'FlowByActivity':
         return FlowByActivity
+
+
+def getFlowByActivity(
+        datasource,
+        year,
+        flowclass=None,
+        geographic_level=None,
+        download_FBA_if_missing=DEFAULT_DOWNLOAD_IF_MISSING
+        ) -> pd.DataFrame:
+    """
+    Retrieves stored data in the FlowByActivity format
+    :param datasource: str, the code of the datasource.
+    :param year: int, a year, e.g. 2012
+    :param flowclass: str or list, a 'Class' of the flow. Optional. E.g.
+    'Water' or ['Employment', 'Chemicals']
+    :param geographic_level: str, a geographic level of the data.
+                             Optional. E.g. 'national', 'state', 'county'.
+    :param download_FBA_if_missing: bool, if True will attempt to load from
+        remote server prior to generating if file not found locally
+    :return: a pandas DataFrame in FlowByActivity format
+    """
+    fba = FlowByActivity.return_FBA(
+        full_name=datasource,
+        config={},
+        year=int(year),
+        download_ok=download_FBA_if_missing
+    )
+
+    if len(fba) == 0:
+        raise flowsa.exceptions.FBANotAvailableError(
+            message=f"Error generating {datasource} for {str(year)}")
+    if flowclass is not None:
+        fba = fba.query('Class == @flowclass')
+    # if geographic level specified, only load rows in geo level
+    if geographic_level is not None:
+        fba = filter_by_geoscale(fba, geographic_level)
+    return pd.DataFrame(fba.reset_index(drop=True))
