+++ scripts/FlowBySector_Activity_Sets/write_FBS_activity_set_BLS_QCEW.py
@@ -8,6 +8,7 @@ run for additional years if there are new NAICS.
 
 import pandas as pd
 import flowsa
+import flowsa.flowbyactivity
 from flowsa.settings import flowbysectoractivitysetspath
 
 datasource = 'BLS_QCEW'
@@ -19,7 +20,7 @@ if __name__ == '__main__':
     # empty df
     df2 = pd.DataFrame()
     for y in as_years:
-        df_import = flowsa.getFlowByActivity(datasource, y)
+        df_import = flowsa.flowbyactivity.getFlowByActivity(datasource, y)
 
         df = (df_import[['ActivityProducedBy']]
               .drop_duplicates()
