+++ scripts/FlowBySector_Activity_Sets/write_FBS_activity_set_EIA_MECS_Land.py
@@ -7,6 +7,7 @@ related to EIA MECS
 """
 
 import flowsa
+import flowsa.flowbyactivity
 from flowsa.settings import flowbysectoractivitysetspath
 
 datasource = 'EIA_MECS_Land'
@@ -14,7 +15,7 @@ datasource = 'EIA_MECS_Land'
 year = '2014'
 
 if __name__ == '__main__':
-    df_import = flowsa.getFlowByActivity(datasource, year)
+    df_import = flowsa.flowbyactivity.getFlowByActivity(datasource, year)
 
     df = (df_import[['ActivityConsumedBy']]
           .drop_duplicates()
